requirements.txt
psycopg2-binary==2.8.6
click==7.1.2
requests==2.25.0